Use floats for positioning

+++ src/ecs/components/position.cpp
@@ -1,25 +1,25 @@
 #include "position.h"
 
 
-Position::Position(double x, double y) {
+Position::Position(float x, float y) {
     m_x = x;
     m_y = y;
 }
 
-double Position::get_x() {
+float Position::get_x() {
     return m_x;
 }
 
-double Position::get_y() {
+float Position::get_y() {
     return m_y;
 }
 
 
-void Position::set_x(double val) {
+void Position::set_x(float val) {
     m_x = val;
 }
 
 
-void Position::set_y(double val) {
+void Position::set_y(float val) {
     m_y = val;
 }
